Add script binning
Add a command line argument to the 02_plant_optimization.py script to enable the processing of only subsets of each given country. This is done to enable the breaking up of computation for large countries (e.g. Norway) into smaller jobs that can then run in parallel. This implementation is currently unelegant and a better solution should be found.