Increase requested wall time
Some jobs terminated because they reached the time limit. The requested wall time was previously 30 hours. This puts the jobs in the 120 hour queue (the next best queue is for <24h jobs). Thus, increasing the requested wall time to any value less than 120h should not deprioritize the jobs to a different queue.